web_server/requirements.txt
Flask==1.1.2
idna==3.3
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==2.11.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==1.1.1
Werkzeug < 1.0.0
pika==1.3.2
requests==2.27.1
python-dotenv



